+++ js/TotalSets.js
@@ -1,7 +1,7 @@
 import Adapt from 'core/js/adapt';
 import Passmark from './Passmark';
 import ScoringSet from './ScoringSet';
-import TotalLifecycleUpdateJournal from './TotalLifecycleUpdateJournal';
+import TotalSetsUpdateJournal from './TotalSetsUpdateJournal';
 import {
   createIntersectedSet
 } from './utils/intersection';
@@ -189,7 +189,7 @@ export default class TotalSets extends ScoringSet {
   /** @override */
   get journal() {
     if (this.isIntersectedSet) return;
-    return (this._journal = this._journal || new TotalLifecycleUpdateJournal({ set: this }));
+    return (this._journal = this._journal || new TotalSetsUpdateJournal({ set: this }));
   }
 
 }
